refactor(autostore): create a separate class for each button
